# Add full seed support