… compat_test.go
… hash128_test.go
… hash64_test.go
… hasher_test.go
… utils.go