Updated version of #4 that duplicates code for seed support to keep speed of unseeded operations.

64 bits ported from #4, and added to Hasher. 128 bits: New code.


```
cpu: AMD Ryzen 9 3950X 16-Core Processor
BenchmarkFixed64/0/default-32         	502115580	         2.427 ns/op
BenchmarkFixed64/0/seed-32            	456687386	         2.626 ns/op
BenchmarkFixed64/1/default-32         	500157549	         2.424 ns/op	 412.57 MB/s
BenchmarkFixed64/1/seed-32            	457582310	         2.624 ns/op	 381.05 MB/s
BenchmarkFixed64/2/default-32         	502988589	         2.515 ns/op	 795.20 MB/s
BenchmarkFixed64/2/seed-32            	455570487	         2.617 ns/op	 764.14 MB/s
BenchmarkFixed64/3/default-32         	556708663	         2.147 ns/op	1397.49 MB/s
BenchmarkFixed64/3/seed-32            	500540583	         2.393 ns/op	1253.45 MB/s
BenchmarkFixed64/4/default-32         	443760067	         2.753 ns/op	1452.91 MB/s
BenchmarkFixed64/4/seed-32            	416873076	         2.869 ns/op	1394.11 MB/s
BenchmarkFixed64/8/default-32         	436751470	         2.743 ns/op	2916.89 MB/s
BenchmarkFixed64/8/seed-32            	416926956	         2.890 ns/op	2768.36 MB/s
BenchmarkFixed64/9/default-32         	498373018	         2.468 ns/op	3646.25 MB/s
BenchmarkFixed64/9/seed-32            	486979585	         2.483 ns/op	3625.13 MB/s
BenchmarkFixed64/16/default-32        	500565638	         2.475 ns/op	6465.45 MB/s
BenchmarkFixed64/16/seed-32           	483982400	         2.469 ns/op	6481.23 MB/s
BenchmarkFixed64/17/default-32        	329918040	         3.621 ns/op	4694.30 MB/s
BenchmarkFixed64/17/seed-32           	304521690	         3.913 ns/op	4344.11 MB/s
BenchmarkFixed64/32/default-32        	324469370	         3.656 ns/op	8752.41 MB/s
BenchmarkFixed64/32/seed-32           	309484858	         3.875 ns/op	8259.06 MB/s
BenchmarkFixed64/33/default-32        	235912114	         5.097 ns/op	6473.79 MB/s
BenchmarkFixed64/33/seed-32           	214626010	         5.583 ns/op	5911.06 MB/s
BenchmarkFixed64/64/default-32        	236769002	         5.067 ns/op	12630.63 MB/s
BenchmarkFixed64/64/seed-32           	216592540	         5.541 ns/op	11549.51 MB/s
BenchmarkFixed64/65/default-32        	178004383	         6.726 ns/op	9663.99 MB/s
BenchmarkFixed64/65/seed-32           	166293452	         7.226 ns/op	8994.99 MB/s
BenchmarkFixed64/96/default-32        	179198703	         6.730 ns/op	14265.35 MB/s
BenchmarkFixed64/96/seed-32           	166370434	         7.256 ns/op	13230.57 MB/s
BenchmarkFixed64/97/default-32        	149042013	         8.057 ns/op	12039.95 MB/s
BenchmarkFixed64/97/seed-32           	136212301	         8.796 ns/op	11027.75 MB/s
BenchmarkFixed64/128/default-32       	149291146	         7.993 ns/op	16014.47 MB/s
BenchmarkFixed64/128/seed-32          	136609206	         8.746 ns/op	14634.74 MB/s
BenchmarkFixed64/129/default-32       	131856292	         9.102 ns/op	14172.10 MB/s
BenchmarkFixed64/129/seed-32          	122928664	         9.730 ns/op	13257.38 MB/s
BenchmarkFixed64/240/default-32       	68028752	        17.42 ns/op	13780.88 MB/s
BenchmarkFixed64/240/seed-32          	62782192	        18.63 ns/op	12879.01 MB/s
BenchmarkFixed64/241-AVX2/default-32  	68358928	        16.55 ns/op	14564.28 MB/s
BenchmarkFixed64/241-AVX2/seed-32     	43491812	        27.32 ns/op	8822.82 MB/s
BenchmarkFixed64/241-SSE2/default-32  	68019112	        17.66 ns/op	13649.86 MB/s
BenchmarkFixed64/241-SSE2/seed-32     	37898766	        31.62 ns/op	7621.44 MB/s
BenchmarkFixed64/241/default-32       	36656107	        32.43 ns/op	7431.37 MB/s
BenchmarkFixed64/241/seed-32          	25395373	        46.93 ns/op	5135.59 MB/s
BenchmarkFixed64/512-AVX2/default-32  	62445034	        19.10 ns/op	26800.54 MB/s
BenchmarkFixed64/512-AVX2/seed-32     	38580122	        30.77 ns/op	16641.55 MB/s
BenchmarkFixed64/512-SSE2/default-32  	42174962	        28.08 ns/op	18230.62 MB/s
BenchmarkFixed64/512-SSE2/seed-32     	29141066	        40.70 ns/op	12580.76 MB/s
BenchmarkFixed64/512/default-32       	21038715	        56.65 ns/op	9037.26 MB/s
BenchmarkFixed64/512/seed-32          	16747822	        71.06 ns/op	7205.48 MB/s
BenchmarkFixed64/1024-AVX2/default-32 	44020704	        27.31 ns/op	37498.26 MB/s
BenchmarkFixed64/1024-AVX2/seed-32    	30545076	        39.12 ns/op	26178.18 MB/s
BenchmarkFixed64/1024-SSE2/default-32 	27156693	        44.69 ns/op	22911.91 MB/s
BenchmarkFixed64/1024-SSE2/seed-32    	20811364	        57.44 ns/op	17827.36 MB/s
BenchmarkFixed64/1024/default-32      	11358676	       105.2 ns/op	9736.26 MB/s
BenchmarkFixed64/1024/seed-32         	10053913	       119.7 ns/op	8554.47 MB/s
BenchmarkFixed64/8192-AVX2/default-32 	 8718560	       136.2 ns/op	60156.17 MB/s
BenchmarkFixed64/8192-AVX2/seed-32    	 8033977	       148.4 ns/op	55212.87 MB/s
BenchmarkFixed64/8192-SSE2/default-32 	 4390557	       274.4 ns/op	29856.16 MB/s
BenchmarkFixed64/8192-SSE2/seed-32    	 4168213	       287.2 ns/op	28525.77 MB/s
BenchmarkFixed64/8192/default-32      	 1513728	       791.9 ns/op	10344.35 MB/s
BenchmarkFixed64/8192/seed-32         	 1436068	       833.9 ns/op	9823.57 MB/s
BenchmarkFixed64/102400-AVX2/default-32         	  744139	      1568 ns/op	65285.90 MB/s
BenchmarkFixed64/102400-AVX2/seed-32            	  743664	      1579 ns/op	64834.99 MB/s
BenchmarkFixed64/102400-SSE2/default-32         	  344264	      3308 ns/op	30957.98 MB/s
BenchmarkFixed64/102400-SSE2/seed-32            	  354777	      3323 ns/op	30816.45 MB/s
BenchmarkFixed64/102400/default-32              	  120934	      9781 ns/op	10469.46 MB/s
BenchmarkFixed64/102400/seed-32                 	  117171	     10171 ns/op	10068.29 MB/s
BenchmarkFixed64/1024000-AVX2/default-32        	   73567	     16082 ns/op	63675.25 MB/s
BenchmarkFixed64/1024000-AVX2/seed-32           	   73556	     16152 ns/op	63396.29 MB/s
BenchmarkFixed64/1024000-SSE2/default-32        	   35470	     33649 ns/op	30431.51 MB/s
BenchmarkFixed64/1024000-SSE2/seed-32           	   35527	     33611 ns/op	30465.87 MB/s
BenchmarkFixed64/1024000/default-32             	   12205	     98762 ns/op	10368.40 MB/s
BenchmarkFixed64/1024000/seed-32                	   10000	    102339 ns/op	10005.99 MB/s
BenchmarkFixed64/10240000-AVX2/default-32       	    6957	    164888 ns/op	62102.84 MB/s
BenchmarkFixed64/10240000-AVX2/seed-32          	    6414	    164867 ns/op	62110.59 MB/s
BenchmarkFixed64/10240000-SSE2/default-32       	    3310	    338310 ns/op	30268.11 MB/s
BenchmarkFixed64/10240000-SSE2/seed-32          	    3445	    351167 ns/op	29159.90 MB/s
BenchmarkFixed64/10240000/default-32            	     946	   1174067 ns/op	8721.82 MB/s
BenchmarkFixed64/10240000/seed-32               	    1120	   1102615 ns/op	9287.01 MB/s
BenchmarkFixed64/102400000-AVX2/default-32      	     210	   5721031 ns/op	17898.87 MB/s
BenchmarkFixed64/102400000-AVX2/seed-32         	     211	   5443095 ns/op	18812.83 MB/s
BenchmarkFixed64/102400000-SSE2/default-32      	     154	   6744777 ns/op	15182.12 MB/s
BenchmarkFixed64/102400000-SSE2/seed-32         	     184	   7197146 ns/op	14227.86 MB/s
BenchmarkFixed64/102400000/default-32           	      91	  13854980 ns/op	7390.84 MB/s
BenchmarkFixed64/102400000/seed-32              	      85	  14555981 ns/op	7034.91 MB/s
```

The Hasher can be reused and keeps the key state.

```
cpu: AMD Ryzen 9 3950X 16-Core Processor            
BenchmarkFixed128/0/default-32         	416331432	         2.850 ns/op
BenchmarkFixed128/0/seed-32            	357644836	         3.362 ns/op
BenchmarkFixed128/1/default-32         	357294813	         3.349 ns/op	 298.64 MB/s
BenchmarkFixed128/1/seed-32            	322657665	         3.718 ns/op	 268.95 MB/s
BenchmarkFixed128/2/default-32         	381831927	         3.159 ns/op	 633.04 MB/s
BenchmarkFixed128/2/seed-32            	327142880	         3.683 ns/op	 543.09 MB/s
BenchmarkFixed128/3/default-32         	388036573	         3.103 ns/op	 966.88 MB/s
BenchmarkFixed128/3/seed-32            	331520743	         3.622 ns/op	 828.29 MB/s
BenchmarkFixed128/4/default-32         	363709435	         3.280 ns/op	1219.38 MB/s
BenchmarkFixed128/4/seed-32            	339963566	         3.528 ns/op	1133.73 MB/s
BenchmarkFixed128/8/default-32         	365782843	         3.281 ns/op	2438.15 MB/s
BenchmarkFixed128/8/seed-32            	339036396	         3.530 ns/op	2266.56 MB/s
BenchmarkFixed128/9/default-32         	289378575	         4.131 ns/op	2178.72 MB/s
BenchmarkFixed128/9/seed-32            	288218145	         4.137 ns/op	2175.43 MB/s
BenchmarkFixed128/16/default-32        	291406282	         4.129 ns/op	3874.64 MB/s
BenchmarkFixed128/16/seed-32           	289152654	         4.153 ns/op	3852.66 MB/s
BenchmarkFixed128/17/default-32        	218846992	         5.412 ns/op	3140.91 MB/s
BenchmarkFixed128/17/seed-32           	207038943	         5.804 ns/op	2929.10 MB/s
BenchmarkFixed128/32/default-32        	223354089	         5.711 ns/op	5603.67 MB/s
BenchmarkFixed128/32/seed-32           	202968138	         5.825 ns/op	5494.02 MB/s
BenchmarkFixed128/33/default-32        	163157976	         7.327 ns/op	4504.06 MB/s
BenchmarkFixed128/33/seed-32           	151520737	         7.901 ns/op	4176.66 MB/s
BenchmarkFixed128/64/default-32        	161572161	         7.369 ns/op	8685.36 MB/s
BenchmarkFixed128/64/seed-32           	152701032	         7.904 ns/op	8096.82 MB/s
BenchmarkFixed128/65/default-32        	129351505	         9.173 ns/op	7086.03 MB/s
BenchmarkFixed128/65/seed-32           	100000000	        10.20 ns/op	6374.84 MB/s
BenchmarkFixed128/96/default-32        	122155568	         9.579 ns/op	10022.25 MB/s
BenchmarkFixed128/96/seed-32           	100000000	        10.59 ns/op	9068.90 MB/s
BenchmarkFixed128/97/default-32        	100000000	        11.47 ns/op	8456.52 MB/s
BenchmarkFixed128/97/seed-32           	100000000	        12.10 ns/op	8014.35 MB/s
BenchmarkFixed128/128/default-32       	100000000	        11.42 ns/op	11212.53 MB/s
BenchmarkFixed128/128/seed-32          	96993210	        12.08 ns/op	10592.00 MB/s
BenchmarkFixed128/129/default-32       	78058934	        15.33 ns/op	8412.91 MB/s
BenchmarkFixed128/129/seed-32          	72235197	        17.75 ns/op	7266.76 MB/s
BenchmarkFixed128/240/default-32       	45628243	        26.87 ns/op	8930.44 MB/s
BenchmarkFixed128/240/seed-32          	41661313	        28.13 ns/op	8531.34 MB/s
BenchmarkFixed128/241-AVX2/default-32  	56973834	        20.71 ns/op	11637.15 MB/s
BenchmarkFixed128/241-AVX2/seed-32     	33875337	        35.48 ns/op	6793.40 MB/s
BenchmarkFixed128/241-SSE2/default-32  	54393646	        22.37 ns/op	10774.23 MB/s
BenchmarkFixed128/241-SSE2/seed-32     	32044263	        37.03 ns/op	6507.65 MB/s
BenchmarkFixed128/241/default-32       	33118796	        36.06 ns/op	6684.07 MB/s
BenchmarkFixed128/241/seed-32          	22310928	        55.60 ns/op	4334.37 MB/s
BenchmarkFixed128/512-AVX2/default-32  	49546034	        23.60 ns/op	21698.48 MB/s
BenchmarkFixed128/512-AVX2/seed-32     	31921005	        37.66 ns/op	13594.64 MB/s
BenchmarkFixed128/512-SSE2/default-32  	37790156	        30.55 ns/op	16761.44 MB/s
BenchmarkFixed128/512-SSE2/seed-32     	25906287	        45.81 ns/op	11177.44 MB/s
BenchmarkFixed128/512/default-32       	20059341	        59.90 ns/op	8548.29 MB/s
BenchmarkFixed128/512/seed-32          	15279169	        80.87 ns/op	6330.89 MB/s
BenchmarkFixed128/1024-AVX2/default-32 	39635616	        30.41 ns/op	33675.21 MB/s
BenchmarkFixed128/1024-AVX2/seed-32    	25689274	        47.70 ns/op	21468.23 MB/s
BenchmarkFixed128/1024-SSE2/default-32 	25274544	        47.07 ns/op	21754.63 MB/s
BenchmarkFixed128/1024-SSE2/seed-32    	19263460	        62.43 ns/op	16401.98 MB/s
BenchmarkFixed128/1024/default-32      	11208228	       107.4 ns/op	9533.96 MB/s
BenchmarkFixed128/1024/seed-32         	 9525070	       126.5 ns/op	8093.44 MB/s
BenchmarkFixed128/8192-AVX2/default-32 	 7391833	       146.8 ns/op	55808.12 MB/s
BenchmarkFixed128/8192-AVX2/seed-32    	 7765632	       155.1 ns/op	52807.87 MB/s
BenchmarkFixed128/8192-SSE2/default-32 	 4311584	       277.4 ns/op	29529.16 MB/s
BenchmarkFixed128/8192-SSE2/seed-32    	 4097803	       293.2 ns/op	27937.89 MB/s
BenchmarkFixed128/8192/default-32      	 1468501	       826.4 ns/op	9912.45 MB/s
BenchmarkFixed128/8192/seed-32         	 1446652	       822.9 ns/op	9955.34 MB/s
BenchmarkFixed128/102400-AVX2/default-32         	  724610	      1575 ns/op	65010.04 MB/s
BenchmarkFixed128/102400-AVX2/seed-32            	  721166	      1586 ns/op	64550.76 MB/s
BenchmarkFixed128/102400-SSE2/default-32         	  356696	      3436 ns/op	29799.67 MB/s
BenchmarkFixed128/102400-SSE2/seed-32            	  357700	      3335 ns/op	30705.46 MB/s
BenchmarkFixed128/102400/default-32              	  118638	     10080 ns/op	10159.16 MB/s
BenchmarkFixed128/102400/seed-32                 	  121681	      9914 ns/op	10328.68 MB/s
BenchmarkFixed128/1024000-AVX2/default-32        	   71001	     16370 ns/op	62555.05 MB/s
BenchmarkFixed128/1024000-AVX2/seed-32           	   62028	     16546 ns/op	61886.37 MB/s
BenchmarkFixed128/1024000-SSE2/default-32        	   33898	     34045 ns/op	30077.97 MB/s
BenchmarkFixed128/1024000-SSE2/seed-32           	   35126	     34051 ns/op	30072.18 MB/s
BenchmarkFixed128/1024000/default-32             	   10000	    102492 ns/op	9991.03 MB/s
BenchmarkFixed128/1024000/seed-32                	   12046	     99191 ns/op	10323.52 MB/s
BenchmarkFixed128/10240000-AVX2/default-32       	    6798	    173702 ns/op	58951.41 MB/s
BenchmarkFixed128/10240000-AVX2/seed-32          	    6829	    176219 ns/op	58109.55 MB/s
BenchmarkFixed128/10240000-SSE2/default-32       	    3202	    350050 ns/op	29253.01 MB/s
BenchmarkFixed128/10240000-SSE2/seed-32          	    3516	    344809 ns/op	29697.61 MB/s
BenchmarkFixed128/10240000/default-32            	    1062	   1080276 ns/op	9479.06 MB/s
BenchmarkFixed128/10240000/seed-32               	    1158	   1048260 ns/op	9768.57 MB/s
BenchmarkFixed128/102400000-AVX2/default-32      	     254	   4547298 ns/op	22518.87 MB/s
BenchmarkFixed128/102400000-AVX2/seed-32         	     260	   4553841 ns/op	22486.51 MB/s
BenchmarkFixed128/102400000-SSE2/default-32      	     178	   6570520 ns/op	15584.76 MB/s
BenchmarkFixed128/102400000-SSE2/seed-32         	     177	   6955960 ns/op	14721.19 MB/s
BenchmarkFixed128/102400000/default-32           	      81	  13826001 ns/op	7406.34 MB/s
BenchmarkFixed128/102400000/seed-32              	      90	  13166462 ns/op	7777.34 MB/s
PASS
ok  	github.com/zeebo/xxh3	121.753s 
```