@klauspost 2021-09-20
Added a copyless preloop for initial writes that are bigger than a block+stride.

Small penalty to small payloads, but great for bigger.

Since Hasher will probably be used mostly for bigger payloads that seems like a reasonable tradeoff.
```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkHasher64/16/go/plain-32               9.07          9.34          +2.95%
BenchmarkHasher64/16/go/seed-32                9.92          10.5          +5.59%
BenchmarkHasher64/16/avx2/plain-32             9.05          9.33          +3.15%
BenchmarkHasher64/16/avx2/seed-32              9.91          10.5          +5.75%
BenchmarkHasher64/16/sse2/plain-32             9.07          9.33          +2.87%
BenchmarkHasher64/16/sse2/seed-32              9.97          10.5          +5.38%
BenchmarkHasher64/64/go/plain-32               12.3          12.7          +3.41%
BenchmarkHasher64/64/go/seed-32                13.2          14.2          +7.28%
BenchmarkHasher64/64/avx2/plain-32             12.2          12.7          +3.84%
BenchmarkHasher64/64/avx2/seed-32              13.3          14.3          +7.45%
BenchmarkHasher64/64/sse2/plain-32             12.2          12.7          +3.92%
BenchmarkHasher64/64/sse2/seed-32              13.2          14.1          +6.83%
BenchmarkHasher64/256/go/plain-32              42.3          44.0          +4.02%
BenchmarkHasher64/256/go/seed-32               58.6          59.9          +2.30%
BenchmarkHasher64/256/avx2/plain-32            27.4          28.3          +3.17%
BenchmarkHasher64/256/avx2/seed-32             40.3          41.2          +2.36%
BenchmarkHasher64/256/sse2/plain-32            27.5          28.4          +3.31%
BenchmarkHasher64/256/sse2/seed-32             42.5          43.9          +3.27%
BenchmarkHasher64/1024/go/plain-32             119           122           +2.45%
BenchmarkHasher64/1024/go/seed-32              137           141           +2.55%
BenchmarkHasher64/1024/avx2/plain-32           42.9          44.5          +3.73%
BenchmarkHasher64/1024/avx2/seed-32            55.9          57.5          +2.88%
BenchmarkHasher64/1024/sse2/plain-32           62.0          64.2          +3.58%
BenchmarkHasher64/1024/sse2/seed-32            73.6          75.7          +2.88%
BenchmarkHasher64/4096/go/plain-32             449           428           -4.79%
BenchmarkHasher64/4096/go/seed-32              458           435           -5.05%
BenchmarkHasher64/4096/avx2/plain-32           127           97.1          -23.72%
BenchmarkHasher64/4096/avx2/seed-32            128           97.0          -24.04%
BenchmarkHasher64/4096/sse2/plain-32           192           166           -13.94%
BenchmarkHasher64/4096/sse2/seed-32            192           166           -13.71%
BenchmarkHasher64/16384/go/plain-32            1797          1643          -8.57%
BenchmarkHasher64/16384/go/seed-32             1835          1679          -8.50%
BenchmarkHasher64/16384/avx2/plain-32          481           304           -36.90%
BenchmarkHasher64/16384/avx2/seed-32           480           304           -36.66%
BenchmarkHasher64/16384/sse2/plain-32          768           596           -22.44%
BenchmarkHasher64/16384/sse2/seed-32           777           595           -23.45%
BenchmarkHasher64/65536/go/plain-32            7509          6521          -13.16%
BenchmarkHasher64/65536/go/seed-32             7651          6601          -13.72%
BenchmarkHasher64/65536/avx2/plain-32          2305          1174          -49.07%
BenchmarkHasher64/65536/avx2/seed-32           2291          1176          -48.67%
BenchmarkHasher64/65536/sse2/plain-32          3404          2306          -32.26%
BenchmarkHasher64/65536/sse2/seed-32           3405          2319          -31.89%
BenchmarkHasher64/262144/go/plain-32           29940         25967         -13.27%
BenchmarkHasher64/262144/go/seed-32            30478         26321         -13.64%
BenchmarkHasher64/262144/avx2/plain-32         9187          4577          -50.18%
BenchmarkHasher64/262144/avx2/seed-32          9180          4592          -49.98%
BenchmarkHasher64/262144/sse2/plain-32         13612         9144          -32.82%
BenchmarkHasher64/262144/sse2/seed-32          13598         9124          -32.90%
BenchmarkHasher64/1048576/go/plain-32          122286        104266        -14.74%
BenchmarkHasher64/1048576/go/seed-32           124692        105724        -15.21%
BenchmarkHasher64/1048576/avx2/plain-32        39926         19192         -51.93%
BenchmarkHasher64/1048576/avx2/seed-32         40002         19250         -51.88%
BenchmarkHasher64/1048576/sse2/plain-32        57867         37980         -34.37%
BenchmarkHasher64/1048576/sse2/seed-32         57824         37972         -34.33%
BenchmarkHasher64/4194304/go/plain-32          489892        436735        -10.85%
BenchmarkHasher64/4194304/go/seed-32           498236        443695        -10.95%
BenchmarkHasher64/4194304/avx2/plain-32        159687        103800        -35.00%
BenchmarkHasher64/4194304/avx2/seed-32         160216        104599        -34.71%
BenchmarkHasher64/4194304/sse2/plain-32        231563        171661        -25.87%
BenchmarkHasher64/4194304/sse2/seed-32         231585        171472        -25.96%
BenchmarkHasher64/16777216/go/plain-32         2419070       2170393       -10.28%
BenchmarkHasher64/16777216/go/seed-32          2614072       2196001       -15.99%
BenchmarkHasher64/16777216/avx2/plain-32       1122912       812622        -27.63%
BenchmarkHasher64/16777216/avx2/seed-32        1030943       821987        -20.27%
BenchmarkHasher64/16777216/sse2/plain-32       1445568       1021819       -29.31%
BenchmarkHasher64/16777216/sse2/seed-32        1433268       1035612       -27.74%
BenchmarkHasher64/67108864/go/plain-32         10717384      8763927       -18.23%
BenchmarkHasher64/67108864/go/seed-32          11764693      8869836       -24.61%
BenchmarkHasher64/67108864/avx2/plain-32       6311313       3413034       -45.92%
BenchmarkHasher64/67108864/avx2/seed-32        6323100       3320728       -47.48%
BenchmarkHasher64/67108864/sse2/plain-32       7547296       4200915       -44.34%
BenchmarkHasher64/67108864/sse2/seed-32        7565688       4158657       -45.03%
BenchmarkHasher64/268435456/go/plain-32        42900148      34811853      -18.85%
BenchmarkHasher64/268435456/go/seed-32         47167288      35460039      -24.82%
BenchmarkHasher64/268435456/avx2/plain-32      25125537      13928190      -44.57%
BenchmarkHasher64/268435456/avx2/seed-32       25088762      13934935      -44.46%
BenchmarkHasher64/268435456/sse2/plain-32      30119150      17303694      -42.55%
BenchmarkHasher64/268435456/sse2/seed-32       30297149      17426832      -42.48%
BenchmarkHasher128/16/go/plain-32              10.8          11.0          +1.75%
BenchmarkHasher128/16/go/seed-32               10.7          11.2          +4.30%
BenchmarkHasher128/16/avx2/plain-32            10.9          11.0          +0.83%
BenchmarkHasher128/16/avx2/seed-32             10.7          10.9          +2.34%
BenchmarkHasher128/16/sse2/plain-32            10.9          11.0          +0.18%
BenchmarkHasher128/16/sse2/seed-32             10.7          10.9          +2.44%
BenchmarkHasher128/64/go/plain-32              14.5          14.8          +1.65%
BenchmarkHasher128/64/go/seed-32               15.1          15.2          +0.93%
BenchmarkHasher128/64/avx2/plain-32            14.5          14.7          +1.73%
BenchmarkHasher128/64/avx2/seed-32             15.1          15.2          +1.00%
BenchmarkHasher128/64/sse2/plain-32            14.5          14.7          +1.73%
BenchmarkHasher128/64/sse2/seed-32             15.1          15.2          +1.13%
BenchmarkHasher128/256/go/plain-32             46.9          62.6          +33.30%
BenchmarkHasher128/256/go/seed-32              64.0          64.4          +0.70%
BenchmarkHasher128/256/avx2/plain-32           31.9          32.5          +1.94%
BenchmarkHasher128/256/avx2/seed-32            52.0          44.4          -14.62%
BenchmarkHasher128/256/sse2/plain-32           32.8          33.1          +1.19%
BenchmarkHasher128/256/sse2/seed-32            47.0          48.0          +2.19%
BenchmarkHasher128/1024/go/plain-32            122           144           +18.07%
BenchmarkHasher128/1024/go/seed-32             143           143           +0.42%
BenchmarkHasher128/1024/avx2/plain-32          48.0          48.4          +0.85%
BenchmarkHasher128/1024/avx2/seed-32           81.5          61.2          -24.86%
BenchmarkHasher128/1024/sse2/plain-32          64.3          65.4          +1.73%
BenchmarkHasher128/1024/sse2/seed-32           78.0          78.5          +0.60%
BenchmarkHasher128/4096/go/plain-32            462           422           -8.59%
BenchmarkHasher128/4096/go/seed-32             461           432           -6.33%
BenchmarkHasher128/4096/avx2/plain-32          130           95.8          -26.46%
BenchmarkHasher128/4096/avx2/seed-32           131           96.3          -26.59%
BenchmarkHasher128/4096/sse2/plain-32          197           176           -10.52%
BenchmarkHasher128/4096/sse2/seed-32           211           180           -14.69%
BenchmarkHasher128/16384/go/plain-32           1823          1639          -10.09%
BenchmarkHasher128/16384/go/seed-32            1840          1677          -8.86%
BenchmarkHasher128/16384/avx2/plain-32         483           306           -36.65%
BenchmarkHasher128/16384/avx2/seed-32          484           304           -37.21%
BenchmarkHasher128/16384/sse2/plain-32         771           603           -21.79%
BenchmarkHasher128/16384/sse2/seed-32          794           607           -23.55%
BenchmarkHasher128/65536/go/plain-32           7543          6483          -14.05%
BenchmarkHasher128/65536/go/seed-32            7638          6588          -13.75%
BenchmarkHasher128/65536/avx2/plain-32         2305          1177          -48.94%
BenchmarkHasher128/65536/avx2/seed-32          2306          1183          -48.70%
BenchmarkHasher128/65536/sse2/plain-32         3429          2318          -32.40%
BenchmarkHasher128/65536/sse2/seed-32          3443          2320          -32.62%
BenchmarkHasher128/262144/go/plain-32          29969         25896         -13.59%
BenchmarkHasher128/262144/go/seed-32           30461         26300         -13.66%
BenchmarkHasher128/262144/avx2/plain-32        9193          4575          -50.23%
BenchmarkHasher128/262144/avx2/seed-32         9221          4575          -50.38%
BenchmarkHasher128/262144/sse2/plain-32        13593         9116          -32.94%
BenchmarkHasher128/262144/sse2/seed-32         13620         9134          -32.94%
BenchmarkHasher128/1048576/go/plain-32         122948        104575        -14.94%
BenchmarkHasher128/1048576/go/seed-32          125890        106158        -15.67%
BenchmarkHasher128/1048576/avx2/plain-32       39954         19196         -51.95%
BenchmarkHasher128/1048576/avx2/seed-32        40129         19288         -51.94%
BenchmarkHasher128/1048576/sse2/plain-32       58118         38131         -34.39%
BenchmarkHasher128/1048576/sse2/seed-32        57932         37946         -34.50%
BenchmarkHasher128/4194304/go/plain-32         493451        416740        -15.55%
BenchmarkHasher128/4194304/go/seed-32          502356        422838        -15.83%
BenchmarkHasher128/4194304/avx2/plain-32       160109        77045         -51.88%
BenchmarkHasher128/4194304/avx2/seed-32        160445        77259         -51.85%
BenchmarkHasher128/4194304/sse2/plain-32       231687        151870        -34.45%
BenchmarkHasher128/4194304/sse2/seed-32        232476        151958        -34.63%
BenchmarkHasher128/16777216/go/plain-32        2418268       1979228       -18.16%
BenchmarkHasher128/16777216/go/seed-32         2594129       2010799       -22.49%
BenchmarkHasher128/16777216/avx2/plain-32      1124619       615349        -45.28%
BenchmarkHasher128/16777216/avx2/seed-32       1028934       601365        -41.55%
BenchmarkHasher128/16777216/sse2/plain-32      1440442       812778        -43.57%
BenchmarkHasher128/16777216/sse2/seed-32       1400430       844471        -39.70%
BenchmarkHasher128/67108864/go/plain-32        10780787      8585564       -20.36%
BenchmarkHasher128/67108864/go/seed-32         11792892      8754885       -25.76%
BenchmarkHasher128/67108864/avx2/plain-32      6311541       3343733       -47.02%
BenchmarkHasher128/67108864/avx2/seed-32       6402239       3303074       -48.41%
BenchmarkHasher128/67108864/sse2/plain-32      7678965       4170337       -45.69%
BenchmarkHasher128/67108864/sse2/seed-32       7689645       4219645       -45.13%
BenchmarkHasher128/268435456/go/plain-32       42878115      34154168      -20.35%
BenchmarkHasher128/268435456/go/seed-32        46528692      34885655      -25.02%
BenchmarkHasher128/268435456/avx2/plain-32     25074578      13406721      -46.53%
BenchmarkHasher128/268435456/avx2/seed-32      24934115      13445259      -46.08%
BenchmarkHasher128/268435456/sse2/plain-32     29948956      16646546      -44.42%
BenchmarkHasher128/268435456/sse2/seed-32      30070987      16912347      -43.76%

benchmark                                      old MB/s     new MB/s     speedup
BenchmarkHasher64/16/go/plain-32               1764.00      1713.51      0.97x
BenchmarkHasher64/16/go/seed-32                1613.58      1528.33      0.95x
BenchmarkHasher64/16/avx2/plain-32             1768.70      1714.76      0.97x
BenchmarkHasher64/16/avx2/seed-32              1614.50      1526.80      0.95x
BenchmarkHasher64/16/sse2/plain-32             1763.91      1714.66      0.97x
BenchmarkHasher64/16/sse2/seed-32              1604.27      1521.86      0.95x
BenchmarkHasher64/64/go/plain-32               5204.80      5029.50      0.97x
BenchmarkHasher64/64/go/seed-32                4853.09      4523.96      0.93x
BenchmarkHasher64/64/avx2/plain-32             5227.29      5034.26      0.96x
BenchmarkHasher64/64/avx2/seed-32              4819.08      4484.34      0.93x
BenchmarkHasher64/64/sse2/plain-32             5226.63      5032.63      0.96x
BenchmarkHasher64/64/sse2/seed-32              4855.94      4544.39      0.94x
BenchmarkHasher64/256/go/plain-32              6047.48      5813.67      0.96x
BenchmarkHasher64/256/go/seed-32               4369.14      4270.74      0.98x
BenchmarkHasher64/256/avx2/plain-32            9337.41      9047.93      0.97x
BenchmarkHasher64/256/avx2/seed-32             6353.36      6207.21      0.98x
BenchmarkHasher64/256/sse2/plain-32            9315.38      9018.63      0.97x
BenchmarkHasher64/256/sse2/seed-32             6018.71      5828.95      0.97x
BenchmarkHasher64/1024/go/plain-32             8635.71      8428.97      0.98x
BenchmarkHasher64/1024/go/seed-32              7454.75      7266.67      0.97x
BenchmarkHasher64/1024/avx2/plain-32           23853.72     22997.97     0.96x
BenchmarkHasher64/1024/avx2/seed-32            18306.73     17793.62     0.97x
BenchmarkHasher64/1024/sse2/plain-32           16516.53     15944.72     0.97x
BenchmarkHasher64/1024/sse2/seed-32            13910.47     13521.23     0.97x
BenchmarkHasher64/4096/go/plain-32             9119.75      9579.02      1.05x
BenchmarkHasher64/4096/go/seed-32              8947.12      9422.36      1.05x
BenchmarkHasher64/4096/avx2/plain-32           32181.53     42184.73     1.31x
BenchmarkHasher64/4096/avx2/seed-32            32074.71     42225.71     1.32x
BenchmarkHasher64/4096/sse2/plain-32           21303.07     24747.53     1.16x
BenchmarkHasher64/4096/sse2/seed-32            21276.46     24666.13     1.16x
BenchmarkHasher64/16384/go/plain-32            9115.98      9972.01      1.09x
BenchmarkHasher64/16384/go/seed-32             8928.75      9757.02      1.09x
BenchmarkHasher64/16384/avx2/plain-32          34062.97     53988.74     1.58x
BenchmarkHasher64/16384/avx2/seed-32           34123.45     53868.80     1.58x
BenchmarkHasher64/16384/sse2/plain-32          21338.09     27515.09     1.29x
BenchmarkHasher64/16384/sse2/seed-32           21087.91     27549.92     1.31x
BenchmarkHasher64/65536/go/plain-32            8727.95      10050.46     1.15x
BenchmarkHasher64/65536/go/seed-32             8566.17      9928.04      1.16x
BenchmarkHasher64/65536/avx2/plain-32          28432.92     55810.43     1.96x
BenchmarkHasher64/65536/avx2/seed-32           28601.55     55723.45     1.95x
BenchmarkHasher64/65536/sse2/plain-32          19249.84     28414.07     1.48x
BenchmarkHasher64/65536/sse2/seed-32           19244.80     28260.91     1.47x
BenchmarkHasher64/262144/go/plain-32           8755.57      10095.42     1.15x
BenchmarkHasher64/262144/go/seed-32            8600.97      9959.38      1.16x
BenchmarkHasher64/262144/avx2/plain-32         28535.51     57275.09     2.01x
BenchmarkHasher64/262144/avx2/seed-32          28554.48     57084.79     2.00x
BenchmarkHasher64/262144/sse2/plain-32         19258.13     28668.38     1.49x
BenchmarkHasher64/262144/sse2/seed-32          19278.77     28731.63     1.49x
BenchmarkHasher64/1048576/go/plain-32          8574.75      10056.76     1.17x
BenchmarkHasher64/1048576/go/seed-32           8409.34      9918.04      1.18x
BenchmarkHasher64/1048576/avx2/plain-32        26263.21     54635.47     2.08x
BenchmarkHasher64/1048576/avx2/seed-32         26213.33     54472.66     2.08x
BenchmarkHasher64/1048576/sse2/plain-32        18120.51     27608.67     1.52x
BenchmarkHasher64/1048576/sse2/seed-32         18133.90     27614.45     1.52x
BenchmarkHasher64/4194304/go/plain-32          8561.68      9603.78      1.12x
BenchmarkHasher64/4194304/go/seed-32           8418.30      9453.12      1.12x
BenchmarkHasher64/4194304/avx2/plain-32        26265.73     40407.47     1.54x
BenchmarkHasher64/4194304/avx2/seed-32         26179.12     40098.91     1.53x
BenchmarkHasher64/4194304/sse2/plain-32        18113.05     24433.71     1.35x
BenchmarkHasher64/4194304/sse2/seed-32         18111.30     24460.51     1.35x
BenchmarkHasher64/16777216/go/plain-32         6935.40      7730.04      1.11x
BenchmarkHasher64/16777216/go/seed-32          6418.04      7639.90      1.19x
BenchmarkHasher64/16777216/avx2/plain-32       14940.81     20645.77     1.38x
BenchmarkHasher64/16777216/avx2/seed-32        16273.67     20410.56     1.25x
BenchmarkHasher64/16777216/sse2/plain-32       11605.96     16418.97     1.41x
BenchmarkHasher64/16777216/sse2/seed-32        11705.57     16200.29     1.38x
BenchmarkHasher64/67108864/go/plain-32         6261.68      7657.40      1.22x
BenchmarkHasher64/67108864/go/seed-32          5704.26      7565.96      1.33x
BenchmarkHasher64/67108864/avx2/plain-32       10633.11     19662.52     1.85x
BenchmarkHasher64/67108864/avx2/seed-32        10613.29     20209.08     1.90x
BenchmarkHasher64/67108864/sse2/plain-32       8891.78      15974.82     1.80x
BenchmarkHasher64/67108864/sse2/seed-32        8870.16      16137.15     1.82x
BenchmarkHasher64/268435456/go/plain-32        6257.22      7711.04      1.23x
BenchmarkHasher64/268435456/go/seed-32         5691.14      7570.08      1.33x
BenchmarkHasher64/268435456/avx2/plain-32      10683.77     19272.82     1.80x
BenchmarkHasher64/268435456/avx2/seed-32       10699.43     19263.49     1.80x
BenchmarkHasher64/268435456/sse2/plain-32      8912.45      15513.19     1.74x
BenchmarkHasher64/268435456/sse2/seed-32       8860.09      15403.57     1.74x
BenchmarkHasher128/16/go/plain-32              1474.02      1449.19      0.98x
BenchmarkHasher128/16/go/seed-32               1495.32      1434.23      0.96x
BenchmarkHasher128/16/avx2/plain-32            1470.89      1457.89      0.99x
…